_dbus_loop_iterate: `timeout` can be defined as int

See merge request dbus/dbus!381

+++ dbus/dbus-mainloop.c
@@ -575,7 +575,7 @@ _dbus_loop_iterate (DBusLoop     *loop,
   DBusList *link;
   int n_ready;
   int initial_serial;
-  long timeout;
+  int timeout;
   int orig_depth;
 
   retval = FALSE;
